requirements.txt
Django==2.2.6
djangorestframework==3.10.3
lxml==4.4.1
python-docx==0.8.10
pytz==2019.3
sqlparse==0.3.0
